// src/features/auth/types/auth.interface.ts
export interface AuthUser {
  id: string
  username: string
  email: string
  role: 'USER' | 'ADMIN'
  avatarUrl: string | null
  bannerUrl: string | null
  bio: string | null
}

export interface AuthSession {
  user: AuthUser
  accessToken: string
}

export interface AuthApiResponse {
  data: AuthSession
}

export interface MessageApiResponse {
  message: string
}

export interface ValidateResetTokenApiResponse {
  data: {
    valid: boolean
  }
}

export interface LoginPayload {
  email: string
  password: string
}

export interface RegisterPayload {
  username: string
  email: string
  password: string
}

export interface ForgotPasswordPayload {
  email: string
}

export interface ResetPasswordPayload {
  token: string
  newPassword: string
}